requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argparse==1.4.0
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cors==1.0.1
django-cors-headers==3.14.0
django-filter==23.1
django-jazzmin==2.6.0
django==4.1.7
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-writable-nested==0.7.0
filelock==3.10.7
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
pillow==9.4.0
psycopg2==2.9.6
pycparser==2.21
pyjwt==2.6.0
pysocks==1.7.1
pytz==2023.3
redis==4.5.4
requests-file==1.5.1
requests==2.28.2
setuptools==67.6.1; python_version >= '3.7'
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tzdata==2023.3
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
